requirements.txt
altair
pyomo
pytessract
google-generativeai
python-dotenv
